UniformGridOpacity: Seperated GL version into own processor
