+++ modules/pathtracing/CMakeLists.txt
@@ -4,16 +4,16 @@ set(HEADER_FILES
     include/inviwo/pathtracing/pathtracingmodule.h
     include/inviwo/pathtracing/pathtracingmoduledefine.h
     include/inviwo/pathtracing/processors/uniformgridopacity.h
+    include/inviwo/pathtracing/processors/uniformgridopacitygl.h
     include/inviwo/pathtracing/processors/volumepathtracer.h
-    #include/inviwo/pathtracing/uniformgrid3d.h
 )
 ivw_group("Header Files" ${HEADER_FILES})
 
 set(SOURCE_FILES
     src/pathtracingmodule.cpp
     src/processors/uniformgridopacity.cpp
+    src/processors/uniformgridopacitygl.cpp
     src/processors/volumepathtracer.cpp
-    # src/uniformgrid3d.cpp
 )
 ivw_group("Source Files" ${SOURCE_FILES})
