+++ modules/pathtracing/src/pathtracingmodule.cpp
@@ -30,6 +30,7 @@
 #include <inviwo/pathtracing/pathtracingmodule.h>
 
 #include <inviwo/pathtracing/processors/uniformgridopacity.h>
+#include <inviwo/pathtracing/processors/uniformgridopacitygl.h>
 #include <inviwo/pathtracing/processors/volumepathtracer.h>
 #include <modules/opengl/shader/shadermanager.h>
 
@@ -44,6 +45,7 @@ PathtracingModule::PathtracingModule(InviwoApplication* app) : InviwoModule(app,
     // Processors
     registerProcessor<VolumePathTracer>();
     registerProcessor<UniformGridOpacity>();
+    registerProcessor<UniformGridOpacityGL>();
 
     // Properties
     // registerProperty<pathtracingProperty>();
